Refactor: Modularize and rename Vanetza integration components
Improved clarity by splitting responsibilities between ApplicationToVanetzaConverter (data preparation) and VanetzaToVeinsBridge (network forwarding).
