+++ src/veins/nodes/Car.ned
@@ -49,12 +49,12 @@ module Car
                 @display("p=130,172;i=block/cogwheel");
         }
 
-        adapter: veins.modules.application.ieee80211p.VanetzaAdapter {
+        bridge: veins.modules.application.ieee80211p.VanetzaToVeinsBridge {
             parameters:
                 @display("p=130,50;i=block/network2");
         }
 
-        provider: veins.modules.application.ieee80211p.VeinsVehicleDataProvider {
+        converter: veins.modules.application.ieee80211p.ApplicationToVanetzaConverter {
             parameters:
                 @display("p=200,50");
         }
@@ -65,13 +65,13 @@ module Car
         // =========================
 
         // before: appl.lowerLayerOut --> nic.upperLayerIn;
-        appl.lowerLayerOut --> adapter.upperLayerIn;
-        adapter.lowerLayerOut --> nic.upperLayerIn;
+        appl.lowerLayerOut --> bridge.upperLayerIn;
+        bridge.lowerLayerOut --> nic.upperLayerIn;
 
 
         // before: nic.upperLayerOut --> appl.lowerLayerIn;
-        nic.upperLayerOut --> adapter.lowerLayerIn;
-        adapter.upperLayerOut --> appl.lowerLayerIn;
+        nic.upperLayerOut --> bridge.lowerLayerIn;
+        bridge.upperLayerOut --> appl.lowerLayerIn;
 
 
         // =========================
@@ -79,13 +79,13 @@ module Car
         // =========================
 
         // before: nic.upperControlIn <-- appl.lowerControlOut;
-        appl.lowerControlOut --> adapter.upperControlIn;
-        adapter.lowerControlOut --> nic.upperControlIn;
+        appl.lowerControlOut --> bridge.upperControlIn;
+        bridge.lowerControlOut --> nic.upperControlIn;
 
 
         // before: nic.upperControlOut --> appl.lowerControlIn;
-        nic.upperControlOut --> adapter.lowerControlIn;
-        adapter.upperControlOut --> appl.lowerControlIn;
+        nic.upperControlOut --> bridge.lowerControlIn;
+        bridge.upperControlOut --> appl.lowerControlIn;
 
 
         // =========================
